requirements.txt
wsgiref==0.1.2
#git+git://github.com/django-nonrel/django-nonrel.git@master
git+git://github.com/django-nonrel/django-permission-backend-nonrel.git@master
#git+https://github.com/adamrpah/mongodb-engine.git
django-filter==0.11.0
#django-mongodb-engine==0.6.0
django-taggit==0.17.0
django-taggit-serializer==0.1.1
djangorestframework==3.1.3
djangorestframework-gis==0.9.3
#djangotoolbox==1.8.0
httpie==0.9.2
Markdown==2.6.2
Pygments==2.0.2
pymongo==3.0.3
requests==2.7.0

git+https://github.com/django-nonrel/django@nonrel-1.6
git+https://github.com/django-nonrel/djangotoolbox
git+https://github.com/django-nonrel/mongodb-engine